Tratando erros 404 utilizando Optional

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -61,24 +62,38 @@ public class TopicosController {
     }
     
     @GetMapping("/{id}")
-    public DetalhesTopicoDto detalhar(@PathVariable Long id) {
-    	Topico topico = topicoRepository.getOne(id);
-    	return new DetalhesTopicoDto(topico);
+    public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
+    	
+    	Optional<Topico> topico = topicoRepository.findById(id);
+    	if(topico.isPresent()) {
+    		return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+    	}
+    	
+    	return ResponseEntity.notFound().build();
     }
     
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-    	Topico topico = form.atualizar(id, topicoRepository);
+    	Optional<Topico> optional = topicoRepository.findById(id);
+    	if(optional.isPresent()) {
+    		Topico topico = form.atualizar(id, topicoRepository);
+    		return ResponseEntity.ok(new TopicoDto(topico));
+    	}
     	
-    	return ResponseEntity.ok(new TopicoDto(topico));
+    	return ResponseEntity.notFound().build();
     }
     
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id) {
-    	topicoRepository.deleteById(id);
-    	return ResponseEntity.noContent().build();
+    	Optional<Topico> topico = topicoRepository.findById(id);
+    	if(topico.isPresent()) {
+    		topicoRepository.deleteById(id);
+    		return ResponseEntity.noContent().build();
+    	}
+    	
+    	return ResponseEntity.notFound().build();
     }
 
 }
